Add Season and Season model updated
