+++ apps/movie/admin.py
@@ -1,6 +1,6 @@
 from django.contrib import admin
-from movie.models import Film, season, View
+from movie.models import Film, Season, View
 # Register your models here.
 admin.site.register(Film)
-admin.site.register(season)
+admin.site.register(Season)
 admin.site.register(View)
